requirements.txt
pandas==2.2.3
scikit-learn==1.6.1
numpy==2.2.3
openai==1.67.0
python-dotenv==1.0.1
tqdm==4.67.1
flask==3.1.3
reportlab==4.4.10
